Add files via upload
Completed the creation of movies to be able to insert it to a BST, or a vector depending on where we go for the project. Hash table is somewhat started, but still has further work to be done for it to actually work. Strongly recommend using BST since it is already implemented from previous projects. 
Movies have a new method called setData() which does all the setters in one go. Also deleted the constructors with multiple parameters as only a default one was created. The setting of data will be done in the business class where buildMovies() helps place the movies in the correct BST.